Refactor Header to Masthead
