+++ src/components/Masthead.js
@@ -3,43 +3,33 @@ import ReactDOM from 'react-dom';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
 import GatsbyImage from './GatsbyImage';
+import { ImageSharpPropTypes } from '../propTypes';
+import { COLORS } from '../styles/vars';
 
 const BackgroundImage = GatsbyImage.extend`
   left: 0;
   top: 0;
   width: 100%;
   height: 100%;
-  opacity: 0.3;
+  opacity: ${props => props.opacity};
 `;
 
-const HeaderContainer = styled.div`
-  margin: 0 auto;
-  max-width: 960px;
-  padding: 1.45rem 1.0875rem;
-  position: relative;
-  z-index: 2;
-  display: flex;
-  justify-content: space-between;
-`;
-
-export class Header extends Component {
-  static displayName = 'Header';
+export class Masthead extends Component {
+  static displayName = 'Masthead';
 
   static propTypes = {
     tag: PropTypes.oneOfType([PropTypes.func, PropTypes.string]),
     children: PropTypes.node,
     isFullHeight: PropTypes.bool,
-    bgImageSizes: PropTypes.shape({
-      src: PropTypes.string.isRequired,
-      base64: PropTypes.string,
-      aspectRatio: PropTypes.number,
-      sizes: PropTypes.string,
-      srcSet: PropTypes.string,
-    }).isRequired,
+    bgImage: ImageSharpPropTypes.isRequired,
+    bgColor: PropTypes.string,
+    opacity: PropTypes.number,
   };
 
   static defaultProps = {
-    tag: 'header',
+    tag: 'div',
+    bgColor: COLORS.bg,
+    opacity: 1,
   };
 
   componentDidUpdate = (prevProps) => {
@@ -63,45 +53,41 @@ export class Header extends Component {
     }
   };
 
+  setRef = (elm) => {
+    // eslint-disable-next-line react/no-find-dom-node
+    this.elm = ReactDOM.findDOMNode(elm);
+  };
+
   render() {
     const {
-      bgImageSizes,
+      bgColor,
+      bgImage,
       isFullHeight,
       children,
+      opacity,
       tag: Tag,
       ...rest
     } = this.props;
+
     return (
-      <Tag
-        ref={(elm) => {
-          // eslint-disable-next-line react/no-find-dom-node
-          this.elm = ReactDOM.findDOMNode(elm);
-        }}
-        {...rest}
-      >
-        <HeaderContainer>
-          {children}
-        </HeaderContainer>
+      <Tag ref={this.setRef} {...rest}>
+        {children}
         <BackgroundImage
           style={{
             position: 'absolute',
           }}
-          sizes={bgImageSizes}
+          sizes={bgImage.sizes}
+          opacity={opacity}
         />
       </Tag>
     );
   }
 }
 
-export default styled(Header)`
-  background: #524763;
+export default styled(Masthead)`
+  background-color: ${props => props.bgColor};
   margin-bottom: 1.45rem;
   overflow: hidden;
   position: relative;
   height: ${({ isFullHeight }) => (isFullHeight ? '70vh' : '20vh')};
-  h1 {
-    img {
-      height: 80px;
-    }
-  }
 `;
