+++ src/components/__tests__/Masthead.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import renderer from 'react-test-renderer';
 import { shallow, mount } from 'enzyme';
 import 'jest-styled-components';
-import Header, { Header as Raw } from '../Header';
+import Masthead, { Masthead as Raw } from '../Masthead';
 
 const props = {
   bgImageSizes: {
@@ -11,7 +11,7 @@ const props = {
 };
 
 it('should render', () => {
-  const component = renderer.create(<Header {...props} />);
+  const component = renderer.create(<Masthead {...props} />);
 
   const tree = component.toJSON();
   expect(tree).toMatchSnapshot();
@@ -35,10 +35,10 @@ it('should display additional classes', () => {
   expect(wrapper.hasClass('my-class')).toBe(true);
 });
 
-it('should render as a <header> by default', () => {
+it('should render as a <div> by default', () => {
   const wrapper = shallow(<Raw {...props} />);
 
-  expect(wrapper.type()).toBe('header');
+  expect(wrapper.type()).toBe('div');
 });
 
 it('should render with a custom tag', () => {
